+++ entity/backend.go
@@ -7,14 +7,20 @@ import (
 
 // nolint: tagliatelle
 type Backend struct {
-	WsConnectionId      string                                    `json:"ws_connection_id"`
-	ModuleId            string                                    `json:"module_id"`
-	Address             string                                    `json:"address"`
-	Version             string                                    `json:"version"`
-	LibVersion          string                                    `json:"lib_version"`
-	ModuleName          string                                    `json:"module_name"`
-	ConfigServiceNodeId string                                    `json:"config_service_node_id"`
-	Endpoints           xtypes.Json[[]cluster.EndpointDescriptor] `json:"endpoints"`
-	RequiredModules     xtypes.Json[[]cluster.ModuleDependency]   `json:"required_modules"`
-	CreatedAt           xtypes.Time                               `json:"created_at"`
+	WsConnectionId       string                                     `json:"ws_connection_id"`
+	ModuleId             string                                     `json:"module_id"`
+	Address              string                                     `json:"address"`
+	Version              string                                     `json:"version"`
+	LibVersion           string                                     `json:"lib_version"`
+	ModuleName           string                                     `json:"module_name"`
+	ConfigServiceNodeId  string                                     `json:"config_service_node_id"`
+	Endpoints            xtypes.Json[[]cluster.EndpointDescriptor]  `json:"endpoints"`
+	RequiredModules      xtypes.Json[[]cluster.ModuleDependency]    `json:"required_modules"`
+	CreatedAt            xtypes.Time                                `json:"created_at"`
+	MetricsAutodiscovery *xtypes.Json[cluster.MetricsAutodiscovery] `json:"metrics_autodiscovery"`
+}
+
+// nolint: tagliatelle
+type MetricsAdWrapper struct {
+	MetricsAutodiscovery xtypes.Json[cluster.MetricsAutodiscovery] `json:"metrics_autodiscovery"`
 }