+++ migrations/20250603131929_add_metrics_autodiscovery_column.sql
@@ -1,0 +1,6 @@
+-- +goose Up
+-- +goose NO TRANSACTION
+alter table isp_config_service__backend add column metrics_autodiscovery blob;
+
+-- +goose Down
+alter table isp_config_service__backend drop column metrics_autodiscovery;